gdr funzionante, montecarlo corretto sono un mostro

+++ esercitazione11_arma/funzioni.h
@@ -39,7 +39,7 @@ double min(double a, double b){
         return b;
     }
 }
-double Fabs(double x){
+double Fabs(int x){
     if(x<0){
         return -x;
     }
